pyproject.toml
# Poetry pyproject.toml: https://python-poetry.org/docs/pyproject/
[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "pycommonknowledge"
version = "0.1.0"
description = "An integrated and opinionated collection of Django applications and Javascript components addressing needs for people building software for organisers and campaigners."
readme = "README.md"
authors = ["Common Knowledge <hello@commonknowledge.coop>"]
license = "GNU GPL v3.0"
repository = "https://github.com/commonknowledge/pycommonknowledge"
homepage = "https://github.com/commonknowledge/pycommonknowledge"

# Keywords description https://python-poetry.org/docs/pyproject/#keywords
keywords = []  #! Update me

# Pypi classifiers: https://pypi.org/classifiers/
classifiers = [  #! Update me
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Operating System :: OS Independent",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
]

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.3.2"}
rich = "^10.7.0"
schedule = "^1.1.0"
djangorestframework-camel-case = "^1.2.0"
djangorestframework-dataclasses = "^1.0.0"

[tool.poetry.dev-dependencies]
django = '^3.2.3'
django-vite = "^1.2.1"
wagtail = '^2.13'
bandit = "^1.7.0"
black = {version = "^21.7b0", allow-prereleases = true}
dj-database-url = '>=0.5.0'
darglint = "^1.8.0"
isort = {extras = ["colors"], version = "^5.9.3"}
mkdocs = "^1.2.3"
mkdocs-material = "^7.3.6"
mypy = "^0.910"
mypy-extensions = "^0.4.3"
pdoc3 = "^0.10.0"
pre-commit = "^2.14.0"
psycopg2 = '>=2.8.6'
pydocstyle = "^6.1.1"
pylint = "^2.10.2"
pytest = "^6.2.4"
pytest-django = "^4.4.0"
pyupgrade = "^2.24.0"
safety = "^1.10.3"

[tool.black]
# https://github.com/psf/black
target-version = ["py39"]
line-length = 88
color = true

exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | node_modules
    | build
    | dist
    | env
    | venv
)/
'''

[tool.isort]
# https://github.com/timothycrosley/isort/
py_version = 39
line_length = 88

known_typing = ["typing", "types", "typing_extensions", "mypy", "mypy_extensions"]
sections = ["FUTURE", "TYPING", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
include_trailing_comma = true
profile = "black"
multi_line_output = 3
indent = 4
color_output = true

[tool.mypy]
# https://mypy.readthedocs.io/en/latest/config_file.html#using-a-pyproject-toml-file
python_version = 3.9
pretty = true
show_traceback = true
color_output = true

allow_redefinition = false
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
ignore_missing_imports = true
implicit_reexport = false
no_implicit_optional = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict_equality = true
strict_optional = true
warn_no_return = false
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true


[tool.pytest.ini_options]
# https://docs.pytest.org/en/6.2.x/customize.html#pyproject-toml
# Directories that are not visited by pytest collector:
norecursedirs =["hooks", "*.egg", ".eggs", "dist", "build", "docs", ".tox", ".git", "__pycache__"]
doctest_optionflags = ["NUMBER", "NORMALIZE_WHITESPACE", "IGNORE_EXCEPTION_DETAIL"]

# Extra options:
addopts = [
  "--strict-markers",
  "--tb=short",
  "--doctest-modules",
  "--doctest-continue-on-failure",
]

DJANGO_SETTINGS_MODULE = "settings"
python_files = "tests.py test_*.py *_tests.py"
